+++ github3/github.py
@@ -616,13 +616,16 @@ class GitHub(GitHubCore):
 
         return self._iter(int(number), url, Organization, etag=etag)
 
-    def iter_repos(self, login=None, type='', sort='', direction='',
-                   number=-1, etag=None):
+    @requires_auth
+    def iter_repos(self, type=None, sort=None, direction=None, number=-1,
+                   etag=None):
         """List public repositories for the specified ``login`` or all
         repositories for the authenticated user if ``login`` is not
         provided.
 
-        :param str login: (optional)
+        .. versionchanged:: 0.6
+           Removed the login parameter for correctness. Use iter_user_repos
+
         :param str type: (optional), accepted values:
             ('all', 'owner', 'public', 'private', 'member')
             API default: 'all'
@@ -639,10 +642,7 @@ class GitHub(GitHubCore):
         :returns: generator of :class:`Repository <github3.repos.Repository>`
             objects
         """
-        if login:
-            url = self._build_url('users', login, 'repos')
-        else:
-            url = self._build_url('user', 'repos')
+        url = self._build_url('user', 'repos')
 
         params = {}
         if type in ('all', 'owner', 'public', 'private', 'member'):
@@ -699,6 +699,41 @@ class GitHub(GitHubCore):
 
         url = self._build_url('user', 'subscriptions')
         return self._iter(int(number), url, Repository, etag=etag)
+
+    def iter_user_repos(self, login, type=None, sort=None, direction=None,
+                        number=-1, etag=None):
+        """List public repositories for the specified ``login``.
+
+        .. versionadded:: 0.6
+
+        :param str login: (required), username
+        :param str type: (optional), accepted values:
+            ('all', 'owner', 'member')
+            API default: 'all'
+        :param str sort: (optional), accepted values:
+            ('created', 'updated', 'pushed', 'full_name')
+            API default: 'created'
+        :param str direction: (optional), accepted values:
+            ('asc', 'desc'), API default: 'asc' when using 'full_name',
+            'desc' otherwise
+        :param int number: (optional), number of repositories to return.
+            Default: -1 returns all repositories
+        :param str etag: (optional), ETag from a previous request to the same
+            endpoint
+        :returns: generator of :class:`Repository <github3.repos.Repository>`
+            objects
+        """
+        url = self._build_url('users', login, 'repos')
+
+        params = {}
+        if type in ('all', 'owner', 'member'):
+            params.update(type=type)
+        if sort in ('created', 'updated', 'pushed', 'full_name'):
+            params.update(sort=sort)
+        if direction in ('asc', 'desc'):
+            params.update(direction=direction)
+
+        return self._iter(int(number), url, Repository, params, etag)
 
     @requires_auth
     def key(self, id_num):