+++ HISTORY.rst
@@ -25,6 +25,10 @@ History/Changelog
 
 - Remove vendored dependency of PySO8601.
 
+- Split `GitHub.iter_repos` into `GitHub.iter_user_repos` and 
+  `GitHub.iter_repos`. As a consequence `github3.iter_repos` is now 
+  `github3.iter_user_repos`
+
 0.5.3: 2013-03-19
 -----------------